requirements.txt
python-telegram-bot==20.x  # Replace with the version you're using
requests